Смена типа на Decimal для количества дней

+++ Interfaces/IPeriod.cs
@@ -30,6 +30,6 @@ namespace CarLeasingViewer
         /// <summary>
         /// Количество дней
         /// </summary>
-        double DayCount { get; }
+        decimal DayCount { get; }
     }
 }
